deformable convoluion v8 class
